+++ src/Handlers/Async_Curl.php
@@ -9,13 +9,6 @@ use Jjbchunta\Async\Handlers\Async_Script;
  * The asynchronous handler responsible for outbound curl requests.
  */
 class Async_Curl extends Async_Script {
-    public static function is_process_of_type( $process ) {
-        if ( !is_string( $process ) ) return false;
-
-        // Evaluate if there is a prefix of an http or https
-        return preg_match( '/^https?:\/\//', $process ) == 1;
-    }
-
     public function __construct( $process, $config = null ) {
         /*
          * With this basically being a wrapper for the `Async_Script` class,