+++ tests/async.php
@@ -25,10 +25,12 @@ echo "Example process completed." . PHP_EOL;
 
 echo "Waiting process finish..." . PHP_EOL;
 $result = await( $promise );
+/*
 echo "Result: " . PHP_EOL;
 echo "```" . PHP_EOL;
 print_r( $result );
 echo PHP_EOL . "```" . PHP_EOL;
+*/
 echo "Process await finished!" . PHP_EOL;
 
 echo "Attempting to run process again..." . PHP_EOL;
@@ -37,9 +39,11 @@ echo "Re-run successfully invocated!" . PHP_EOL;
 
 echo "Waiting process finish..." . PHP_EOL;
 $result = await( $promise );
+/*
 echo "Result: " . PHP_EOL;
 echo "```" . PHP_EOL;
 print_r( $result );
 echo "```" . PHP_EOL;
+*/
 echo "Process await finished!" . PHP_EOL;
 exit;